requirements.txt
pandas
pytest
pytest-runner
python-binance
wheel
setuptools
twine